Move platform guessing to __init__.py
And import appropriate platform-specific module automatically
